// StageV/src/app/services/models/visa-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Passport } from '../models/passport';
export interface VisaRequest {
  countryIssuingVisa?: string;
  countryOfFirstMission?: string;
  dateOfMission?: string;
  expiryDate?: string;
  id?: number;
  issueDate?: string;
  issuedBy?: string;
  issuingCountry?: string;
  missionPurpose?: string;
  passport?: Passport;
  passportNumber?: string;
  status?: string;
  travelerType?: string;
}